Do a little more style cleanup.
