+++ src/System.Windows.Forms.Primitives/tests/UnitTests/Interop/Ole32/CADWORDTests.cs
@@ -52,7 +52,7 @@ namespace System.Windows.Forms.Primitives.Ole32Tests
         }
 
         private static Ole32.CADWORD CreateIntVector(params uint[] values)
-            => CreateIntVector(null, values);
+            => CreateIntVector(allocations: null, values);
 
         private unsafe static Ole32.CADWORD CreateIntVector(IList<IntPtr>? allocations, params uint[] values)
         {
